[Work] Opdata Download Test Complete ★★
- m_CurrentSaveIndex 의 Type 을 unsigned short 에서 int 로 바꿔주었음.. 이녀석 때문에 저장이 제대로 안 되는 것이었음...

+++ SourceCode/Main.h
@@ -134,7 +134,7 @@ public: // Download Functions
 	bool __fastcall SendRequestPacket(unsigned char _FCode, unsigned short _StartIdx, unsigned short _Size);
     unsigned short m_StartIdx;
     unsigned short m_TotalDataBlockCount;
-    unsigned short m_CurrentSaveIndex;
+    int m_CurrentSaveIndex;
     unsigned short __fastcall GetCRCValue(BYTE* _pData, int _DataSize);
 	BYTE m_FileBuf[FILE_SAVE_BUF_SIZE];
     bool __fastcall SaveFile(int _Type);
